+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Tests for "Network cabling".
 - Tests for "Scrabble".
+- Tests for "Stock exchange losses".
 
 ## [1.2.0] - 2022-03-02
 ### Added