+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Scrabble".
 - Tests for "Stock exchange losses".
 - Tests for "Telephone numbers".
+- Tests for "The gift".
 
 ## [1.2.0] - 2022-03-02
 ### Added