+++ CHANGELOG.md
@@ -4,13 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.3.0] - 2022-03-02
 ### Added
 - Tests for "Network cabling".
 - Tests for "Scrabble".
 - Tests for "Stock exchange losses".
 - Tests for "Telephone numbers".
 - Tests for "The gift".
+- Tests for "War".
 
 ## [1.2.0] - 2022-03-02
 ### Added